+++ core/src/main/java/org/lemurproject/galago/core/index/corpus/DocumentReader.java
@@ -17,7 +17,7 @@ public interface DocumentReader extends IndexPartReader {
 
   public abstract Document getDocument(byte[] key, DocumentComponents p) throws IOException;
 
-  public abstract Document getDocument(int key, DocumentComponents p) throws IOException;
+  public abstract Document getDocument(long key, DocumentComponents p) throws IOException;
 
   public interface DocumentIterator extends KeyIterator {
