+++ core/src/main/java/org/lemurproject/galago/core/retrieval/iterator/ExtentInsideIterator.java
@@ -54,7 +54,7 @@ public class ExtentInsideIterator extends ExtentConjunctionIterator {
   @Override
   public void loadExtents() {
     // get the document
-    int document = context.document;
+    long document = context.document;
 
     // check if we're already there
     if (context.cachable && this.extentCache.getDocument() == document) {
