+++ core/src/main/java/org/lemurproject/galago/core/retrieval/iterator/OrderedWindowIterator.java
@@ -23,7 +23,7 @@ public class OrderedWindowIterator extends ExtentConjunctionIterator {
   @Override
   public void loadExtents() {
     // get the document
-    int document = context.document;
+    long document = context.document;
 
     // check if we're already there
     if (context.cachable && this.extentCache.getDocument() == document) {
