+++ core/src/main/java/org/lemurproject/galago/core/retrieval/iterator/ScoreCombinationIterator.java
@@ -94,7 +94,7 @@ public class ScoreCombinationIterator extends DisjunctionIterator implements Sco
     String type = "score";
     String className = this.getClass().getSimpleName();
     String parameters = np.toString();
-    int document = currentCandidate();
+    long document = currentCandidate();
     boolean atCandidate = hasMatch(this.context.document);
     String returnValue = Double.toString(score());
     List<AnnotatedNode> children = new ArrayList();
