+++ core/src/main/java/org/lemurproject/galago/core/retrieval/iterator/ScoringFunctionIterator.java
@@ -71,7 +71,7 @@ public class ScoringFunctionIterator extends TransformIterator implements ScoreI
     String type = "score";
     String className = this.getClass().getSimpleName();
     String parameters = np.toString();
-    int document = currentCandidate();
+    long document = currentCandidate();
     boolean atCandidate = hasMatch(this.context.document);
     String returnValue = Double.toString(score());
     List<AnnotatedNode> children = new ArrayList();
