+++ core/src/main/java/org/lemurproject/galago/core/retrieval/iterator/SynonymIterator.java
@@ -22,7 +22,7 @@ public class SynonymIterator extends ExtentDisjunctionIterator {
 
   public void loadExtents() {
     // get the document
-    int document = context.document;
+    long document = context.document;
 
     // check if we're already there
     if (context.cachable && this.extentCache.getDocument() == document) {
