+++ core/src/main/java/org/lemurproject/galago/core/retrieval/iterator/ThresholdIterator.java
@@ -50,7 +50,7 @@ public class ThresholdIterator extends TransformIterator implements IndicatorIte
     String type = "score";
     String className = this.getClass().getSimpleName();
     String parameters = "";
-    int document = currentCandidate();
+    long document = currentCandidate();
     boolean atCandidate = hasMatch(this.context.document);
     String returnValue = Boolean.toString(indicator(this.context.document));
     List<AnnotatedNode> children = Collections.singletonList(this.iterator.getAnnotatedNode());
